#443
- popup: обновлено позиционирование в соответствии с новыми требованиями (patch)
- popup: декомпозиция stories (patch)
- deps: обновлен пакет `@floating-ui/react` (patch)
