+++ package.json
@@ -29,7 +29,7 @@
     "storybook:build": "BROWSERSLIST_ENV=\"storybook\" npx storybook build --quiet --loglevel error -o build-storybook"
   },
   "dependencies": {
-    "@floating-ui/react": "^0.25.4",
+    "@floating-ui/react": "^0.26.13",
     "@krutoo/input-mask": "^0.0.9",
     "body-scroll-lock": "3.1.5",
     "classnames": "^2.3.2",
