+++ src/popup/__stories__/01-primary.stories.tsx
@@ -1,0 +1,19 @@
+import { PopupView } from '@sima-land/ui-nucleons/popup';
+
+export default {
+  title: 'common/Popup',
+  component: PopupView,
+  parameters: {
+    layout: 'padded',
+  },
+};
+
+export function Primary() {
+  return (
+    <>
+      <PopupView onClose={() => alert('Нажат крестик!')}>Это тестовый popup.</PopupView>
+    </>
+  );
+}
+
+Primary.storyName = 'Простой пример';
